TASK41-extract Viewer into own package
