+++ view/__init__.py
@@ -1,0 +1,2 @@
+from .base_classes import *
+from .rectangle_packer import *
